// src/domain/enums/SpecializationType.ts
export enum SpecializationType {
  PERSONAL_TRAINING = 'PERSONAL_TRAINING',
  YOGA = 'YOGA',
  PILATES = 'PILATES',
  MARTIAL_ARTS = 'MARTIAL_ARTS',
  CROSSFIT = 'CROSSFIT',
  SWIMMING = 'SWIMMING',
  NUTRITION_COACHING = 'NUTRITION_COACHING',
  DANCE = 'DANCE',
  CYCLING = 'CYCLING',
  REHABILITATION = 'REHABILITATION',
  MEDITATION = 'MEDITATION',
  FUNCTIONAL_TRAINING = 'FUNCTIONAL_TRAINING',
}